backend/foodgram/requirements.txt
asgiref==3.5.2
Django==3.2.15
django-colorfield==0.7.2
djangorestframework==3.14.0
djoser==2.1.0
Pillow==9.2.0
psycopg2-binary==2.8.6
python-dotenv==0.21.0
pytz==2022.2.1
simplejwt==4.7.2
sqlparse==0.4.2
